--- app/src/main/java/com/example/moviews/data/remote/ParseDataWithJson.kt
@@ -1,66 +1,0 @@
-package com.example.moviews.data.remote
-
-import com.example.moviews.data.model.Genre
-import com.example.moviews.data.model.Movie
-import com.example.moviews.utils.Constant
-import org.json.JSONException
-import org.json.JSONObject
-import java.io.BufferedReader
-import java.io.InputStreamReader
-import java.net.HttpURLConnection
-import java.net.URL
-
-class ParseDataWithJson {
-
-    fun getJsonFromUrl(urlString: String?): String {
-        val url = URL(urlString)
-        val httpURLConnection = url.openConnection() as HttpURLConnection
-        httpURLConnection.apply {
-            connectTimeout = Constant.BASE_TIME_OUT
-            readTimeout = Constant.BASE_TIME_OUT
-            requestMethod = Constant.BASE_METHOD_GET
-            doOutput = true
-            connect()
-        }
-
-        val bufferedReader = BufferedReader(InputStreamReader(url.openStream()))
-        val stringBuilder = StringBuilder()
-        var line: String?
-        while (bufferedReader.readLine().also { line = it } != null) {
-            stringBuilder.append(line)
-        }
-        bufferedReader.close()
-        httpURLConnection.disconnect()
-        return stringBuilder.toString()
-    }
-
-    fun parseJsonToData(jsonObject: JSONObject?, keyEntity: String): Any {
-        val data = mutableListOf<Any>()
-        try {
-            val jsonArray = jsonObject?.getJSONArray(keyEntity)
-            for (i in 0 until (jsonArray?.length() ?: 0)) {
-                val jsonObjects = jsonArray?.getJSONObject(i)
-                val item = parseJsonToObject(jsonObjects, keyEntity)
-                item?.let { data.add(it) }
-            }
-        } catch (e: JSONException) {
-            e.printStackTrace()
-        }
-        return data
-    }
-
-    private fun parseJsonToObject(jsonObject: JSONObject?, keyEntity: String): Any? {
-        try {
-            jsonObject?.let {
-                when (keyEntity) {
-                    Movie.MOVIE_RESULTS -> return Movie(it)
-                    Genre.GENRES -> return Genre(it)
-                    else -> null
-                }
-            }
-        } catch (e: JSONException) {
-            e.printStackTrace()
-        }
-        return null
-    }
-}
